Board id 수정

+++ src/main/kotlin/com/b2/backoffice/domain/board/dto/BoardResponse.kt
@@ -3,7 +3,7 @@ package com.b2.backoffice.domain.board.dto
 import java.time.LocalDateTime
 
 data class BoardResponse(
-    val id : Long, // 자동 생성
+    val id : Int, // 자동 생성
     var createAt : LocalDateTime,
     var title : String,
     var contents : String,
